+++ pkg/contexts/ocm/repositories/genericocireg/componentversion.go
@@ -9,6 +9,7 @@ import (
 	"path"
 	"strings"
 
+	"github.com/open-component-model/ocm/pkg/contexts/ocm/cpi/repocpi"
 	"github.com/opencontainers/go-digest"
 
 	"github.com/open-component-model/ocm/pkg/common"
@@ -27,7 +28,6 @@ import (
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/compdesc"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/cpi"
 	"github.com/open-component-model/ocm/pkg/contexts/ocm/cpi/accspeccpi"
-	"github.com/open-component-model/ocm/pkg/contexts/ocm/cpi/support"
 	"github.com/open-component-model/ocm/pkg/errors"
 	"github.com/open-component-model/ocm/pkg/generics"
 	"github.com/open-component-model/ocm/pkg/refmgmt"
@@ -39,18 +39,18 @@ func newComponentVersionAccess(mode accessobj.AccessMode, comp *componentAccessI
 	if err != nil {
 		return nil, err
 	}
-	impl, err := support.NewComponentVersionAccessImpl(comp.GetName(), version, c, true, persistent, !compositionmodeattr.Get(comp.GetContext()))
+	impl, err := repocpi.NewComponentVersionAccessBase(comp.GetName(), version, c, true, persistent, !compositionmodeattr.Get(comp.GetContext()))
 	if err != nil {
 		c.Close()
 		return nil, err
 	}
-	return cpi.NewComponentVersionAccess(impl), nil
+	return repocpi.NewComponentVersionAccess(impl), nil
 }
 
 // //////////////////////////////////////////////////////////////////////////////
 
 type ComponentVersionContainer struct {
-	impl support.ComponentVersionAccessImpl
+	impl repocpi.ComponentVersionAccessBase
 
 	comp     *componentAccessImpl
 	version  string
@@ -59,7 +59,7 @@ type ComponentVersionContainer struct {
 	state    accessobj.State
 }
 
-var _ support.ComponentVersionContainer = (*ComponentVersionContainer)(nil)
+var _ repocpi.ComponentVersionAccessImpl = (*ComponentVersionContainer)(nil)
 
 func newComponentVersionContainer(mode accessobj.AccessMode, comp *componentAccessImpl, version string, access oci.ArtifactAccess) (*ComponentVersionContainer, error) {
 	m := access.ManifestAccess()
@@ -81,11 +81,11 @@ func newComponentVersionContainer(mode accessobj.AccessMode, comp *componentAcce
 	}, nil
 }
 
-func (c *ComponentVersionContainer) SetImplementation(impl support.ComponentVersionAccessImpl) {
+func (c *ComponentVersionContainer) SetImplementation(impl repocpi.ComponentVersionAccessBase) {
 	c.impl = impl
 }
 
-func (c *ComponentVersionContainer) GetParentViewManager() cpi.ComponentAccessViewManager {
+func (c *ComponentVersionContainer) GetParentViewManager() repocpi.ComponentAccessViewManager {
 	return c.comp
 }
