+++ pkg/refmgmt/refmgmt.go
@@ -14,7 +14,7 @@ import (
 
 var ALLOC_REALM = logging.DefineSubRealm("reference counting", "refcnt")
 
-var allocLog = logging.DynamicLogger(ALLOC_REALM)
+var AllocLog = logging.DynamicLogger(ALLOC_REALM)
 
 type Allocatable interface {
 	Ref() error
@@ -83,7 +83,7 @@ func (c *refMgmt) Ref() error {
 		return ErrClosed
 	}
 	c.refcount++
-	allocLog.Trace("ref", "name", c.name, "refcnt", c.refcount)
+	AllocLog.Trace("ref", "name", c.name, "refcnt", c.refcount)
 	return nil
 }
 
@@ -97,7 +97,7 @@ func (c *refMgmt) Unref() error {
 	var err error
 
 	c.refcount--
-	allocLog.Trace("unref", "name", c.name, "refcnt", c.refcount)
+	AllocLog.Trace("unref", "name", c.name, "refcnt", c.refcount)
 	if c.refcount <= 0 {
 		for _, f := range c.before {
 			f.Cleanup()
@@ -135,13 +135,14 @@ func (c *refMgmt) UnrefLast() error {
 	}
 
 	if c.refcount > 1 {
+		AllocLog.Trace("unref last failed", "name", c.name, "pending", c.refcount)
 		return errors.ErrStillInUseWrap(errors.Newf("%d reference(s) pending", c.refcount), c.name)
 	}
 
 	var err error
 
 	c.refcount--
-	allocLog.Trace("unref last", "name", c.name, "refcnt", c.refcount)
+	AllocLog.Trace("unref last", "name", c.name, "refcnt", c.refcount)
 	if c.refcount <= 0 {
 		for _, f := range c.before {
 			f.Cleanup()
@@ -154,7 +155,7 @@ func (c *refMgmt) UnrefLast() error {
 	}
 
 	if err != nil {
-		allocLog.Trace("cleanup last failed", "name", c.name, "error", err.Error())
+		AllocLog.Trace("cleanup last failed", "name", c.name, "error", err.Error())
 		return errors.Wrapf(err, "unable to cleanup %s while unref last", c.name)
 	}
